final_pjt_back/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.1
idna==3.7
requests==2.31.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
pillow==11.0.0